requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.34.2
PyYAML==6.0.2
libretranslatepy==2.1.1
numpy==1.26.4
requests==2.31.0


